requirements.txt
anyio==3.6.1
click==8.1.3
dnspython==2.2.1
fastapi==0.81.0
h11==0.13.0
idna==3.3
pydantic==1.10.1
pymongo==4.2.0
python-dotenv==0.20.0
sniffio==1.3.0
starlette==0.19.1
typing-extensions==4.3.0
uvicorn==0.18.3
anyio==3.6.1
click==8.1.3
dnspython==2.2.1
fastapi==0.81.0
h11==0.13.0
idna==3.3
pydantic==1.10.1
pylint-venv==2.3.0
pymongo==4.2.0
python-dotenv==0.20.0
sniffio==1.3.0
starlette==0.19.1
typing-extensions==4.3.0
uvicorn==0.18.3
